requirements.txt
flask
torch
torchvision
timm
opencv-python
numpy
pillow
